requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
boto3==1.34.22
botocore==1.34.22
certifi==2023.11.17
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.0
fastdup==1.85
h11==0.14.0
idna==3.6
ijson==3.2.3
imageio==2.33.1
importlib-metadata==7.0.1
ipykernel==6.29.0
ipython==8.18.1
jedi==0.19.1
jmespath==1.0.1
jupyter-client==8.6.0
jupyter-core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.9
numpy==1.26.3
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pillow-heif==0.14.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic-core==2.14.6
pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.28.1
s3transfer==0.10.0
sentry-sdk==1.39.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.35.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing-extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.26.0
wcwidth==0.2.13
zipp==3.17.0
